+++ native/src/lib.rs
@@ -6,7 +6,7 @@ use napi_derive::napi;
 mod archive;
 mod archive_gzip;
 mod task;
-#[cfg(target_os = "linux")]
+#[cfg(any(target_os = "linux", target_os = "macos", windows))]
 mod watch;
 mod fast_file;
 mod clone_tree;